BF: replace vcrpy with pytest-vcr; create new test casettes
There where a lot of problems caused by the usage of that
specific version of vcrpy, which blocked a lot of other efforts
to uplift the tests.

This commit replaces it with pytest-vcr, and updates rereferences
to it accordingly in all places (README.md, pyproject.toml, tests).

This change allowed to remove a bunch of other code that was just
needed to make vcrpy work.

The old casettes were replaced since they where out of date.
To create new cassettes, I had to:
- delete old casettes
- locally overwrite the tests/fixtures/dummy_config.json with a valid config
- start the tests so they do real "requests" once and record them
- add test username/password/addresses consistently:
    - replace the dummy_config.json with test values
    - change the information in the casettes to match